+++ src/utils/ray_utils.c
@@ -6,13 +6,13 @@
 /*   By: msavelie <msavelie@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 11:25:02 by msavelie          #+#    #+#             */
-/*   Updated: 2025/04/14 15:27:04 by msavelie         ###   ########.fr       */
+/*   Updated: 2025/05/12 15:21:05 by msavelie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/mini_rt.h"
 
-inline double	length2(t_vector vec)
+inline float	length2(t_vector vec)
 {
 	return (pow(vec.x, 2) + pow(vec.y, 2) + pow(vec.z, 2));
 }
@@ -22,12 +22,12 @@ inline t_vector	revert_vector(t_vector vec)
 	return ((t_vector) {-vec.x, -vec.y, -vec.z});
 }
 
-inline double	dot(t_vector vec1, t_vector vec2)
+inline float	dot(t_vector vec1, t_vector vec2)
 {
 	return (vec1.x * vec2.x + vec1.y * vec2.y + vec1.z * vec2.z);
 }
 
-inline double	lerp(double a, double b, double mix)
+inline float	lerp(float a, float b, float mix)
 {
 	return (b * mix + a * (1 - mix));
 }
