+++ src/utils/vec_utils.c
@@ -6,7 +6,7 @@
 /*   By: msavelie <msavelie@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 16:23:59 by msavelie          #+#    #+#             */
-/*   Updated: 2025/04/23 16:33:24 by msavelie         ###   ########.fr       */
+/*   Updated: 2025/05/12 15:21:07 by msavelie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,12 +27,12 @@ inline t_vector vec_mul(t_vector vec1, t_vector vec2)
 	return ((t_vector) { vec1.x * vec2.x, vec1.y * vec2.y, vec1.z * vec2.z });
 }
 
-inline t_vector vec_mul_num(t_vector vec1, double num)
+inline t_vector vec_mul_num(t_vector vec1, float num)
 {
 	return ((t_vector) { vec1.x * num, vec1.y * num, vec1.z * num });
 }
 
-inline t_vector vec_sub_num(t_vector vec1, double num)
+inline t_vector vec_sub_num(t_vector vec1, float num)
 {
 	return ((t_vector) { vec1.x - num, vec1.y - num, vec1.z - num });
 }
